fix: ObserverOngoingRoom/DebateChatObserverMode 컴포넌트 분리
